+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+# 2.0.4 Pre-Release:
+## Update Notes:
+- CPR for v14 does not contain all automations that previously existed in V13.
+- Generic Macros: `{type}Bonus` and `reroll{Type}WithBonus`, where type is check, save, skill, or tool.
+## Bug Fixes:
+- Generic: `movementAnimation`
+- Instinctive Pounce
+- Cunning Action
+## Updated Legacy Automations:
+- Indomitable
+- Second Wind
+## Updated Modern Automations:
+- Indomitable
+- Second Wind
+- Studied Attacks
+- Tactical Mind
+- Tactical Shift
+
 # 2.0.3 Pre-Release:
 ## Update Notes:
 - CPR for v14 does not contain all automations that previously existed in V13.